+++ src/check/args.ts
@@ -1,4 +1,4 @@
-// generated by src/check/check.schema.ts
+// generated by src/check/args.schema.ts
 
 export interface CheckTaskArgs {
 	typecheck?: boolean;
@@ -28,4 +28,4 @@ export interface CheckTaskArgs {
 	'no-lint'?: boolean;
 }
 
-// generated by src/check/check.schema.ts
+// generated by src/check/args.schema.ts
